refactor(api): C5 — config/profiles/settings-transfer service modules (ADR-0015)
- NEW api/config_service.py: apply_config_patch(body, config=, dev_types=,
  managers=, reload=) — the PUT body verbatim; the config object is patched
  via per-field setattr, NEVER rebound (identity is shared by managers and
  runtimes); rollback-by-restore on reload failure unchanged.
- NEW api/profiles_service.py: list/get/save/apply/rename/delete bodies +
  require_no_active_runs + snapshot_warnings. apply receives shared_breakers
  and forge_runtime.breakers per call (monkeypatch-friendly).
- NEW api/settings_transfer.py: _PassphraseRequired, open_uploaded_bundle,
  export/import/env bodies with explicit deps.
- main.py: 13 routes become 1-4 line forwards with unchanged coroutine
  names/signatures — test_profiles/test_settings_transfer pass untouched
  except ONE monkeypatch retarget (MAX_BUNDLE_BYTES now lives in the
  service module).

Verified: ruff clean; full suite 637 passed / 1 skipped via pytest_app.sh.
